@@ -2,74 +2,113 @@ let React = require('react');
 let d3 = require('d3');
 
 let TooltipMixin = {
-	propTypes: {
-		tooltipHtml: React.PropTypes.func
-	},
+    propTypes: {
+        tooltipHtml: React.PropTypes.func,
+        tooltipMode: React.PropTypes.oneOf(['mouse', 'element', 'fixed']),
+        tooltipContained: React.PropTypes.bool,
+        tooltipOffset: React.PropTypes.objectOf(React.PropTypes.number)
+    },
 
-	getInitialState() {
-		return {
-			tooltip: {
-				hidden: true
-			}
-		};
-	},
+    getInitialState() {
+        return {
+            tooltip: {
+                hidden: true
+            }
+        };
+    },
 
-	getDefaultProps() {
-		return {
-			tooltipOffset: {top: -20, left: 15},
-			tooltipHtml: null
-		};
-	},
+    getDefaultProps() {
+        return {
+            tooltipMode: 'mouse',
+            tooltipOffset: {top: -35, left: 0},
+            tooltipHtml: null,
+            tooltipContained: false
+        };
+    },
 
-	componentDidMount() {
-		this._svg_node = this.getDOMNode().getElementsByTagName("svg")[0];
-	},
+    componentDidMount() {
+        this._svg_node = this.getDOMNode().getElementsByTagName("svg")[0];
+    },
 
-	onMouseEnter(e, data) {
-		if (!this.props.tooltipHtml) {
-			return;
-		}
+    onMouseEnter(e, data) {
+        if (!this.props.tooltipHtml) {
+            return;
+        }
 
-		e.preventDefault();
+        e.preventDefault();
 
-		let {margin, tooltipHtml} = this.props;
+        let {margin,
+             tooltipMode,
+             tooltipOffset,
+             tooltipContained} = this.props;
 
-		let svg = this._svg_node;
-		let position;
-		if (svg.createSVGPoint) {
-			var point = svg.createSVGPoint();
-			point.x = e.clientX, point.y = e.clientY;
-			point = point.matrixTransform(svg.getScreenCTM().inverse());
-			position = [point.x - margin.left, point.y - margin.top];
-		} else {
-			let rect = svg.getBoundingClientRect();
-			position = [e.clientX - rect.left - svg.clientLeft - margin.left,
-						e.clientY - rect.top - svg.clientTop - margin.top];
-		}
+        let svg = this._svg_node;
+        let position;
+        if (svg.createSVGPoint) {
+            var point = svg.createSVGPoint();
+            point.x = e.clientX, point.y = e.clientY;
+            point = point.matrixTransform(svg.getScreenCTM().inverse());
+            position = [point.x - margin.left, point.y - margin.top];
+        } else {
+            let rect = svg.getBoundingClientRect();
+            position = [e.clientX - rect.left - svg.clientLeft - margin.left,
+                        e.clientY - rect.top - svg.clientTop - margin.top];
+        }
 
-		this.setState({
-			tooltip: {
-				top: e.clientY + this.props.tooltipOffset.top,
-				left: e.clientX + this.props.tooltipOffset.left,
-				hidden: false,
-				html: this._tooltipHtml(data, position)
-			}
-		});
-	},
+        let [html, xPos, yPos] = this._tooltipHtml(data, position);
 
-	onMouseLeave(e) {
-		if (!this.props.tooltipHtml) {
-			return;
-		}
+        let svgTop = svg.getBoundingClientRect().top + margin.top;
+        let svgLeft = svg.getBoundingClientRect().left + margin.left;
 
-		e.preventDefault();
+        let top = 0;
+        let left = 0;
 
-		this.setState({
-			tooltip: {
-				hidden: true
-			}
-		});
-	}
+        if (tooltipMode === 'fixed') {
+            top = svgTop + tooltipOffset.top;
+            left = svgLeft + tooltipOffset.left;
+        } else if (tooltipMode === 'element') {
+            top = svgTop + yPos + tooltipOffset.top;
+            left = svgLeft + xPos + tooltipOffset.left;
+        } else { // mouse
+            top = e.clientY + tooltipOffset.top;
+            left = e.clientX + tooltipOffset.left;
+        }
+
+        function lerp(t, a, b) {
+            return (1 - t) * a + t * b;
+        }
+
+        let translate = 50;
+
+        if (tooltipContained) {
+            let t = position[0] / svg.getBoundingClientRect().width;
+            translate = lerp(t, 0, 100);
+        }
+
+        this.setState({
+            tooltip: {
+                top: top,
+                left: left,
+                hidden: false,
+                html: html,
+                translate: translate
+            }
+        });
+    },
+
+    onMouseLeave(e) {
+        if (!this.props.tooltipHtml) {
+            return;
+        }
+
+        e.preventDefault();
+
+        this.setState({
+            tooltip: {
+                hidden: true
+            }
+        });
+    }
 };
 
 module.exports = TooltipMixin;
